create maven project from scaffold and set configuration; update README
