+++ README.md
@@ -1,48 +1,13 @@
-# Introduction
+# MacheteV2
 
-The Dropwizard example application was developed to, as its name implies, provide examples of some of the features
-present in Dropwizard.
+How to start the MacheteV2 application
+---
 
-# Overview
+1. Run `mvn clean install` to build your application
+1. Start application with `java -jar target/machete-1.0-SNAPSHOT.jar server config.yml`
+1. To check that your application is running enter url `http://localhost:8080`
 
-Included with this application is an example of the optional DB API module. The examples provided illustrate a few of
-the features available in [Hibernate](http://hibernate.org/), along with demonstrating how these are used from within
-Dropwizard.
+Health Check
+---
 
-This database example is comprised of the following classes:
-
-* The `PersonDAO` illustrates using the Data Access Object pattern with assisting of Hibernate.
-
-* The `Person` illustrates mapping of Java classes to database tables with assisting of JPA annotations.
-
-* All the JPQL statements for use in the `PersonDAO` are located in the `Person` class.
-
-* `migrations.xml` illustrates the usage of `dropwizard-migrations` which can create your database prior to running
-your application for the first time.
-
-* The `PersonResource` and `PeopleResource` are the REST resource which use the PersonDAO to retrieve data from the database, note the injection
-of the PersonDAO in their constructors.
-
-As with all the modules the db example is wired up in the `initialize` function of the `HelloWorldApplication`.
-
-# Running The Application
-
-To test the example application run the following commands.
-
-* To setup the h2 database run.
-
-        java -jar target/dropwizard-example-$DW_VERSION.jar db migrate example.yml
-
-* To run the server run.
-
-        java -jar target/dropwizard-example-$DW_VERSION.jar server example.yml
-
-* To hit the Hello World example (hit refresh a few times).
-
-	http://localhost:8080/hello-world
-
-* To post data into the application.
-
-	curl -H "Content-Type: application/json" -X POST -d '{"fullName":"Other Person","jobTitle":"Other Title"}' http://localhost:8080/people
-	
-	open http://localhost:8080/people
+To see your applications health enter url `http://localhost:8081/healthcheck`
