+++ machete.yml
@@ -1,0 +1,4 @@
+logging:
+  level: INFO
+  loggers:
+    machete: DEBUG
