--- src/main/java/com/example/helloworld/filter/DateRequired.java
@@ -1,11 +1,0 @@
-package com.example.helloworld.filter;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface DateRequired {
-}
