--- src/main/java/com/example/helloworld/filter/DateRequiredFeature.java
@@ -1,16 +1,0 @@
-package com.example.helloworld.filter;
-
-import javax.ws.rs.container.DynamicFeature;
-import javax.ws.rs.container.ResourceInfo;
-import javax.ws.rs.core.FeatureContext;
-import javax.ws.rs.ext.Provider;
-
-@Provider
-public class DateRequiredFeature implements DynamicFeature {
-    @Override
-    public void configure(ResourceInfo resourceInfo, FeatureContext context) {
-        if (resourceInfo.getResourceMethod().getAnnotation(DateRequired.class) != null) {
-            context.register(DateNotSpecifiedFilter.class);
-        }
-    }
-}
