--- src/main/java/com/example/helloworld/resources/FilteredResource.java
@@ -1,17 +1,0 @@
-package com.example.helloworld.resources;
-
-import com.example.helloworld.filter.DateRequired;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-
-@Path("/filtered")
-public class FilteredResource {
-
-    @GET
-    @DateRequired
-    @Path("hello")
-    public String sayHello() {
-        return "hello";
-    }
-}
