--- src/main/java/com/example/helloworld/resources/PeopleResource.java
@@ -1,37 +1,0 @@
-package com.example.helloworld.resources;
-
-import com.example.helloworld.core.Person;
-import com.example.helloworld.db.PersonDAO;
-import io.dropwizard.hibernate.UnitOfWork;
-
-import javax.validation.Valid;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-import java.util.List;
-
-@Path("/people")
-@Produces(MediaType.APPLICATION_JSON)
-public class PeopleResource {
-
-    private final PersonDAO peopleDAO;
-
-    public PeopleResource(PersonDAO peopleDAO) {
-        this.peopleDAO = peopleDAO;
-    }
-
-    @POST
-    @UnitOfWork
-    public Person createPerson(@Valid Person person) {
-        return peopleDAO.create(person);
-    }
-
-    @GET
-    @UnitOfWork
-    public List<Person> listPeople() {
-        return peopleDAO.findAll();
-    }
-
-}
