--- src/main/java/com/example/helloworld/views/PersonView.java
@@ -1,32 +1,0 @@
-package com.example.helloworld.views;
-
-import com.example.helloworld.core.Person;
-import io.dropwizard.views.View;
-
-public class PersonView extends View {
-    private final Person person;
-
-    public enum Template {
-        FREEMARKER("freemarker/person.ftl"),
-        MUSTACHE("mustache/person.mustache");
-
-        private String templateName;
-
-        Template(String templateName) {
-            this.templateName = templateName;
-        }
-
-        public String getTemplateName() {
-            return templateName;
-        }
-    }
-
-    public PersonView(PersonView.Template template, Person person) {
-        super(template.getTemplateName());
-        this.person = person;
-    }
-
-    public Person getPerson() {
-        return person;
-    }
-}
