--- src/main/resources/assets/example.txt
@@ -1,1 +1,0 @@
-Hello, I'm an example static asset file.
