+++ src/main/resources/banner.txt
@@ -1,8 +1,6 @@
-                           web-scale hello world dP for the web
-                                                 88
-  .d8888b. dP.  .dP .d8888b. 88d8b.d8b. 88d888b. 88 .d8888b.
-  88ooood8  `8bd8'  88'  `88 88'`88'`88 88'  `88 88 88ooood8
-  88.  ...  .d88b.  88.  .88 88  88  88 88.  .88 88 88.  ...
-  `88888P' dP'  `dP `88888P8 dP  dP  dP 88Y888P' dP `88888P'
-                                        88
-                                        dP
+================================================================================
+
+                              MacheteV2
+
+================================================================================
+
