--- src/main/resources/com/example/helloworld/views/mustache/person.mustache
@@ -1,9 +1,0 @@
-<html>
-    <head>
-      <link rel="stylesheet" href="/assets/pure-min.css">
-    </head>
-    <body>
-        <h1>Hello, {{person.fullName}}!</h1>
-        You are an awesome {{person.jobTitle}}!
-    </body>
-</html>
