--- src/main/resources/views/ftl/iso88591.ftl
@@ -1,10 +1,0 @@
-<html>
-<body>
-
-<h1>This is an example of a view containing ISO-8859-1 characters</h1>
-
-¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢
-
-</body>
-</html>
-
