--- src/main/resources/views/mustache/iso88591.mustache
@@ -1,10 +1,0 @@
-<html>
-<body>
-
-<h1>This is an example of a view containing ISO-8859-1 characters</h1>
-
-¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢¢
-
-</body>
-</html>
-
