--- src/main/resources/views/mustache/utf8.mustache
@@ -1,9 +1,0 @@
-<html>
-<body>
-
-<h1>This is an example of a view containing UTF-8 characters</h1>
-
-€€€€€€€€€€€€€€€€€€
-
-</body>
-</html>
