--- src/test/java/com/example/helloworld/IntegrationTest.java
@@ -1,104 +1,0 @@
-package com.example.helloworld;
-
-import com.example.helloworld.api.Saying;
-import com.example.helloworld.core.Person;
-import io.dropwizard.testing.ConfigOverride;
-import io.dropwizard.testing.ResourceHelpers;
-import io.dropwizard.testing.junit5.DropwizardAppExtension;
-import io.dropwizard.testing.junit5.DropwizardExtensionsSupport;
-import org.eclipse.jetty.http.HttpStatus;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-import javax.ws.rs.client.Entity;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Optional;
-
-import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.assertj.core.api.Assertions.assertThat;
-
-@ExtendWith(DropwizardExtensionsSupport.class)
-public class IntegrationTest {
-
-    private static final String TMP_FILE = createTempFile();
-    private static final String CONFIG_PATH = ResourceHelpers.resourceFilePath("test-example.yml");
-
-    public static final DropwizardAppExtension<HelloWorldConfiguration> RULE = new DropwizardAppExtension<>(
-            HelloWorldApplication.class, CONFIG_PATH,
-            ConfigOverride.config("database.url", "jdbc:h2:" + TMP_FILE));
-
-    @BeforeAll
-    public static void migrateDb() throws Exception {
-        RULE.getApplication().run("db", "migrate", CONFIG_PATH);
-    }
-
-    private static String createTempFile() {
-        try {
-            return File.createTempFile("test-example", null).getAbsolutePath();
-        } catch (IOException e) {
-            throw new IllegalStateException(e);
-        }
-    }
-
-    @Test
-    public void testHelloWorld() throws Exception {
-        final Optional<String> name = Optional.of("Dr. IntegrationTest");
-        final Saying saying = RULE.client().target("http://localhost:" + RULE.getLocalPort() + "/hello-world")
-                .queryParam("name", name.get())
-                .request()
-                .get(Saying.class);
-        assertThat(saying.getContent()).isEqualTo(RULE.getConfiguration().buildTemplate().render(name));
-    }
-
-    @Test
-    public void testPostPerson() throws Exception {
-        final Person person = new Person("Dr. IntegrationTest", "Chief Wizard", 1525);
-        final Person newPerson = postPerson(person);
-        assertThat(newPerson.getId()).isNotNull();
-        assertThat(newPerson.getFullName()).isEqualTo(person.getFullName());
-        assertThat(newPerson.getJobTitle()).isEqualTo(person.getJobTitle());
-    }
-
-    @Test
-    public void testRenderingPersonFreemarker() throws Exception {
-        testRenderingPerson("view_freemarker");
-    }
-
-    @Test
-    public void testRenderingPersonMustache() throws Exception {
-        testRenderingPerson("view_mustache");
-    }
-
-    private void testRenderingPerson(String viewName) throws Exception {
-        final Person person = new Person("Dr. IntegrationTest", "Chief Wizard", 1525);
-        final Person newPerson = postPerson(person);
-        final String url = "http://localhost:" + RULE.getLocalPort() + "/people/" + newPerson.getId() + "/" + viewName;
-        Response response = RULE.client().target(url).request().get();
-        assertThat(response.getStatus()).isEqualTo(HttpStatus.OK_200);
-    }
-
-    private Person postPerson(Person person) {
-        return RULE.client().target("http://localhost:" + RULE.getLocalPort() + "/people")
-                .request()
-                .post(Entity.entity(person, MediaType.APPLICATION_JSON_TYPE))
-                .readEntity(Person.class);
-    }
-
-    @Test
-    public void testLogFileWritten() throws IOException {
-        // The log file is using a size and time based policy, which used to silently
-        // fail (and not write to a log file). This test ensures not only that the
-        // log file exists, but also contains the log line that jetty prints on startup
-        final Path log = Paths.get("./logs/application.log");
-        assertThat(log).exists();
-        final String actual = new String(Files.readAllBytes(log), UTF_8);
-        assertThat(actual).contains("0.0.0.0:" + RULE.getLocalPort());
-    }
-}
