--- src/test/resources/test-docker-example.yml
@@ -1,29 +1,0 @@
-template: Hello, %s!
-defaultName: Stranger
-
-database:
-  driverClass: org.postgresql.Driver
-  user: postgres
-  password: postgres
-  url: jdbc:postgresql://localhost:5432/example
-
-server:
-  applicationConnectors:
-    - type: http
-      port: 0
-  adminConnectors:
-    - type: http
-      port: 0
-
-# Logging settings.
-logging:
-  level: INFO
-  appenders:
-    - type: console
-    - type: file
-      currentLogFilename: './logs/application.log'
-      archivedLogFilenamePattern: './logs/application-%d-%i.log.gz'
-      archive: true
-      archivedFileCount: 7
-      maxFileSize: '1MiB'
-
